additional int overflow preventin in v_compute_3d

+++ src/v_compute_3d.hh
@@ -62,7 +62,7 @@ class voro_compute_3d {
         const int hxy;
         /** A constant, set to the value of hx*hy*hz, which is used in the
          * routines which step through mask boxes in sequence. */
-        const int hxyz;
+        const uint64_t hxyz;
         /** The number of floating point entries to store for each particle. */
         const int ps;
         /** This array holds the numerical IDs of each particle in each
@@ -100,7 +100,7 @@ class voro_compute_3d {
         double *mrad;
         /** This array is used during the cell computation to determine which
          * blocks have been considered. */
-        unsigned int *mask;
+        uint64_t *mask;
         /** An array is used to store the queue of blocks to test during the
          * Voronoi cell computation. */
         int *qu;
@@ -124,13 +124,13 @@ class voro_compute_3d {
         bool compute_min_max_radius(int di,int dj,int dk,double fx,double fy,double fz,double gx,double gy,double gz,double& crs,double mrs,double &r_mul);
         bool compute_min_radius(int di,int dj,int dk,double fx,double fy,double fz,double mrs);
         inline void add_to_mask(int ei,int ej,int ek,int *&qu_e);
-        inline void scan_bits_mask_add(unsigned int q,unsigned int *mijk,int ei,int ej,int ek,int *&qu_e);
+        inline void scan_bits_mask_add(unsigned int q,uint64_t *mijk,int ei,int ej,int ek,int *&qu_e);
         inline void scan_all(int ijk,double x,double y,double z,int di,int dj,int dk,particle_record_3d &w,double &mrs);
         void add_list_memory(int*& qu_s,int*& qu_e);
         /** Resets the mask in cases where the mask counter wraps
          * around. */
         inline void reset_mask() {
-            for(unsigned int *mp(mask);mp<mask+hxyz;mp++) *mp=0;
+            for(uint64_t *mp(mask);mp<mask+hxyz;mp++) *mp=0;
         }
 };
